removed repeated code and clean up
