+++ src/lib/components/todo-item.svelte
